+++ engine/docs/changelog.md
@@ -1,6 +1,9 @@
 # TBD
-### Features
+### Fixes
 * Added conditions to handle not having api container's public IP and ports when getting enclave response from the backend
+
+### Changes
+* Upgraded to container-engine-lib 0.20.2, which add some backend methods
 
 # 1.18.1
 ### Features